[refactor] Use camelCase for internal typescript files
