+++ src/internal/fungibleAsset.ts
